+++ CHANGELOG.md
@@ -11,5 +11,6 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - `SerialConnectionManager` to maintain serial connection
 - `SerialReader` along with `SerialPacket` (and descendants) to manage serial-data
 - `OverlayWindow` to show climate control status
+- `RemoteControl` to handle steering wheel button events
 
 [Unreleased]: https://github.com/rampage128/cardroid